feat: Refactor amount handling and error handling
- Added import statement for `com.metaplex.amount.SolAmount` in "Accounts.kt"
- Updated dependency in `solana_public_keys` module to `"project(mapOf("path" to ":amount"))"` in "build.gradle.kts"
- Created "Errors.kt" file with sealed class "AmountErrors" and subclasses "UnexpectedAmountError" and "AmountMismatchError" in "amount/src/commonMain/kotlin/com/metaplex/amount" directory
- Renamed and moved file `solana_interfaces/src/commonMain/kotlin/com/metaplex/solana_interfaces/Amount.kt` to `amount/src/commonMain/kotlin/com/metaplex/amount/Amount.kt`
- Removed unused imports and commented code, renamed `UnexpectedAmountError` to `AmountError`, and used `AmountErrors` in "SolanaError.kt"
- Added `:amount` module to the project in "settings.gradle.kts"
- Created `amount/build.gradle.kts` file with plugins, dependencies, and configurations for `com.ionspin.kotlin:bignum:0.3.8` and `kotlin("test")`
- Renamed and updated package import in "AmountTests.kt", added tests and utils for amount operations and formatting
