+++ settings.gradle.kts
@@ -21,3 +21,4 @@ include(":base58")
 include(":solana_eddsa")
 include(":solana_interfaces")
 include(":solana_keypair")
+include(":amount")
